src/main/java/com/creditunion/service/AccountService.java
package com.creditunion.service;

import com.creditunion.repository.AccountRepository;
import lombok.RequiredArgsConstructor;
import com.creditunion.model.Account;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
@RequiredArgsConstructor
public class AccountService {

    private final AccountRepository accountRepository;

    public List<Account> getAllAccounts() {
        return accountRepository.findAll();
    }

    public Account getAccountById(Long id) {
        return accountRepository.findById(id).orElseThrow(() -> new RuntimeException("Account not found"));
    }

    public Account createAccount(Account account) {
        return accountRepository.save(account);
    }

    public Account updateAccount(Long id, Account accountDetails) {
        Account account = accountRepository.findById(id).orElseThrow(() -> new RuntimeException("Account not found"));
        account.setFirstName(accountDetails.getFirstName());
        account.setLastName(accountDetails.getLastName());
        account.setAddress(accountDetails.getAddress());
        return accountRepository.save(account);
    }

    public void deleteAccount(Long id) {
        accountRepository.deleteById(id);
    }

    public List<Account> searchAccounts(String query) {
        // Implement your search logic here, e.g., using JPA query methods
        return accountRepository.findByFirstNameContainingOrLastNameContaining(query, query);
    }
}

